% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headerEWMA.R
\name{getCCEWMACond}
\alias{getCCEWMACond}
\title{Add together two numbers}
\usage{
getCCEWMACond(
  p0,
  interval = c(1, 4),
  ARL0,
  eps,
  lambda,
  m,
  nu = m - 1,
  n = 5,
  ubc = 1,
  t.interval = c(20, 50),
  state = "steady",
  tol = 1e-04
)
}
\arguments{

\item{p0}{A number}

\item{interval}{A number}

\item{ARL0}{A number}

\item{eps}{A number}

\item{lambda}{A number}

\item{m}{A number}

\item{nu}{A number}

\item{n}{A number}

\item{ubc}{A number}

\item{t.interval}{A number}

\item{state}{A number}

\item{tol}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
getCCEWMACond(p0 = 0.05, interval = c(1, 5), ARL0 = 370, eps = 0, lambda = 0.1, m = 50, nu = 49, 
	n = 5, ubc = 1, tmin = 20, tmax = 50, state = 'steady', tol = 1e-2)
}
