% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headerEWMA.R
\name{getCCEWMAUncond}
\alias{getCCEWMAUncond}
\title{Add together two numbers}
\usage{
getCCEWMAUncond(
  ARL0,
  interval,
  lambda,
  m,
  nu,
  n,
  ubc,
  tmin,
  tmax,
  state = "steady",
  tol = 0.01
)
}
\arguments{
\item{ARL0}{A number}

\item{interval}{A number}

\item{lambda}{A number}

\item{m}{A number}

\item{nu}{A number}

\item{n}{A number}

\item{ubc}{A number}

\item{tmin}{A number}

\item{tmax}{A number}

\item{state}{A number}

\item{tol}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
getCCEWMAUncond(ARL0 = 370, interval = c(1, 5), lambda = 0.1, m = 50, nu = 49, 
	n = 5, ubc = 1, tmin = 20, tmax = 50, state = 'steady', tol = 1e-2)
}
